+++ docs/examples/e06_simple_bot_structure/main.py
@@ -1,0 +1,21 @@
+from dotenv import dotenv_values
+from hata import Client
+from hata.ext.extension_loader import EXTENSION_LOADER
+
+config = dotenv_values('.env')
+
+Sakuya = Client(config['TOKEN'])
+
+@Sakuya.events
+async def ready(client):
+    print(f'{client:f} is connected!')
+
+# Registered variables to extension loader, show up in each loaded extension file.
+EXTENSION_LOADER.add_default_variables(Sakuya=Sakuya)
+# Adds the extensions file or the extensions files recursive in the folder
+EXTENSION_LOADER.add('modules')
+# Loads all the added extension files.
+EXTENSION_LOADER.load_all()
+
+
+Sakuya.start()